Normal-approximation-based test for two-sample BF problem proposed by Srivastava et al. (2013)

Srivastava et al. (2013)'s test for testing equality of two-sample high-dimensional mean vectors without assuming that two covariance matrices are the same.

Usage

SKK2013.TSBF.NABT(y1, y2)

Arguments

y1

The data matrix (\(n_1 \times p\)) from the first population. Each row represents a \(p\)-dimensional observation.

y2

The data matrix (\(n_2 \times p\) from the second population. Each row represents a \(p\)-dimensional observation.

Value

A list of class "NRtest" containing the results of the hypothesis test. See the help file for NRtest.object for details.

Details

Suppose we have two independent high-dimensional samples: $$ \boldsymbol{y}_{i1},\ldots,\boldsymbol{y}_{in_i}, \;\operatorname{are \; i.i.d. \; with}\; \operatorname{E}(\boldsymbol{y}_{i1})=\boldsymbol{\mu}_i,\; \operatorname{Cov}(\boldsymbol{y}_{i1})=\boldsymbol{\Sigma}_i,i=1,2. $$ The primary object is to test $$H_{0}: \boldsymbol{\mu}_1 = \boldsymbol{\mu}_2\; \operatorname{versus}\; H_{1}: \boldsymbol{\mu}_1 \neq \boldsymbol{\mu}_2.$$ Srivastava et al. (2013) proposed the following test statistic: $$T_{SKK} = \frac{(\bar{\boldsymbol{y}}_1 - \bar{\boldsymbol{y}}_2)^\top \hat{\boldsymbol{D}}^{-1}(\bar{\boldsymbol{y}}_1 - \bar{\boldsymbol{y}}_2) - p}{\sqrt{2 \widehat{\operatorname{Var}}(\hat{q}_n) c_{p,n}}},$$ where \(\bar{\boldsymbol{y}}_{i},i=1,2\) are the sample mean vectors, \(\hat{\boldsymbol{D}}=\hat{\boldsymbol{D}}_1/n_1+\hat{\boldsymbol{D}}_2/n_2\) with \(\hat{\boldsymbol{D}}_i,i=1,2\) being the diagonal matrices consisting of only the diagonal elements of the sample covariance matrices. \(\widehat{\operatorname{Var}}(\hat{q}_n)\) is given by equation (1.18) in Srivastava et al. (2013), and \(c_{p, n}\) is the adjustment coefficient proposed by Srivastava et al. (2013). They showed that under the null hypothesis, \(T_{SKK}\) is asymptotically normally distributed.

References

Srivastava_2013HDNRA

Examples

library("HDNRA")
data("COVID19")
dim(COVID19)
#> [1]    87 20460
group1 <- as.matrix(COVID19[c(2:19, 82:87), ]) ## healthy group
group2 <- as.matrix(COVID19[-c(1:19, 82:87), ]) ## COVID-19 patients
SKK2013.TSBF.NABT(group1,group2)
#> 
#> Results of Hypothesis Test
#> --------------------------
#> 
#> Test name:                       Srivastava et al. (2013)'s test
#> 
#> Null Hypothesis:                 Difference between two mean vectors is 0
#> 
#> Alternative Hypothesis:          Difference between two mean vectors is not 0
#> 
#> Data:                            group1 and group2
#> 
#> Sample Sizes:                    n1 = 24
#>                                  n2 = 62
#> 
#> Sample Dimension:                20460
#> 
#> Test Statistic:                  T[SKK] = 2.8966
#> 
#> Approximation method to the      Normal approximation
#> null distribution of T[SKK]: 
#> 
#> Approximation parameter(s):      Adjustment coefficient = 17.9488
#> 
#> P-value:                         0.001886357
#>