Normal-reference-test with three-cumulant (3-c) matched χ^2-approximation for GLHT problem under heteroscedasticity proposed by Zhang and Zhu (2022)

Zhang and Zhu (2022)'s test for general linear hypothesis testing (GLHT) problem for high-dimensional data under heteroscedasticity.

Usage

ZZ2022.GLHTBF.3cNRT(Y,G,n,p)

Arguments

Y

A list of \(k\) data matrices. The \(i\)th element represents the data matrix (\(n_i\times p\)) from the \(i\)th population with each row representing a \(p\)-dimensional observation.

G

A known full-rank coefficient matrix (\(q\times k\)) with \(\operatorname{rank}(\boldsymbol{G})< k\).

n

A vector of \(k\) sample sizes. The \(i\)th element represents the sample size of group \(i\), \(n_i\).

p

The dimension of data.

Value

A list of class "NRtest" containing the results of the hypothesis test. See the help file for NRtest.object for details.

Details

Suppose we have the following \(k\) independent high-dimensional samples: $$ \boldsymbol{y}_{i1},\ldots,\boldsymbol{y}_{in_i}, \;\operatorname{are \; i.i.d. \; with}\; \operatorname{E}(\boldsymbol{y}_{i1})=\boldsymbol{\mu}_i,\; \operatorname{Cov}(\boldsymbol{y}_{i1})=\boldsymbol{\Sigma}_i,i=1,\ldots,k. $$ It is of interest to test the following GLHT problem: $$H_0: \boldsymbol{G M}=\boldsymbol{0}, \quad \text { vs. } H_1: \boldsymbol{G M} \neq \boldsymbol{0},$$ where \(\boldsymbol{M}=(\boldsymbol{\mu}_1,\ldots,\boldsymbol{\mu}_k)^\top\) is a \(k\times p\) matrix collecting \(k\) mean vectors and \(\boldsymbol{G}:q\times k\) is a known full-rank coefficient matrix with \(\operatorname{rank}(\boldsymbol{G})<k\).

Let \(\bar{\boldsymbol{y}}_{i},i=1,\ldots,k\) be the sample mean vectors and \(\hat{\boldsymbol{\Sigma}}_i,i=1,\ldots,k\) be the sample covariance matrices.

Zhang and Zhu (2022) proposed the following U-statistic based test statistic: $$ T_{ZZ}=\|\boldsymbol{C \hat{\mu}}\|^2-\sum_{i=1}^kh_{ii}\operatorname{tr}(\hat{\boldsymbol{\Sigma}}_i)/n_i, $$ where \(\boldsymbol{C}=[(\boldsymbol{G D G}^\top)^{-1/2}\boldsymbol{G}]\otimes\boldsymbol{I}_p\), \(\boldsymbol{D}=\operatorname{diag}(1/n_1,\ldots,1/n_k)\), and \(h_{ij}\) is the \((i,j)\)th entry of the \(k\times k\) matrix \(\boldsymbol{H}=\boldsymbol{G}^\top(\boldsymbol{G}\boldsymbol{D}\boldsymbol{G}^\top)^{-1}\boldsymbol{G}\).

References

Zhang_2022heteroscedasticHDNRA

Examples

library("HDNRA")
data("corneal")
dim(corneal)
#> [1]  150 2000
group1 <- as.matrix(corneal[1:43, ]) ## normal group
group2 <- as.matrix(corneal[44:57, ]) ## unilateral suspect group
group3 <- as.matrix(corneal[58:78, ]) ## suspect map group
group4 <- as.matrix(corneal[79:150, ]) ## clinical keratoconus group
p <- dim(corneal)[2]
Y <- list()
k <- 4
Y[[1]] <- group1
Y[[2]] <- group2
Y[[3]] <- group3
Y[[4]] <- group4
n <- c(nrow(Y[[1]]),nrow(Y[[2]]),nrow(Y[[3]]),nrow(Y[[4]]))
G <- cbind(diag(k-1),rep(-1,k-1))
ZZ2022.GLHTBF.3cNRT(Y,G,n,p)
#> 
#> Results of Hypothesis Test
#> --------------------------
#> 
#> Test name:                       Zhang and Zhu (2022)'s test
#> 
#> Null Hypothesis:                 The general linear hypothesis is true
#> 
#> Alternative Hypothesis:          The general linear hypothesis is not true
#> 
#> Data:                            Y
#> 
#> Sample Sizes:                    n1 = 43
#>                                  n2 = 14
#>                                  n3 = 21
#>                                  n4 = 72
#> 
#> Sample Dimension:                2000
#> 
#> Test Statistic:                  T[ZZ] = 121.1988
#> 
#> Approximation method to the      3-c matched chi^2-approximation
#> null distribution of T[ZZ]: 
#> 
#> Approximation parameter(s):      df    =   4.9334
#>                                  beta0 = -35.0606
#>                                  beta1 =   7.1068
#> 
#> P-value:                         0.0004959474
#>