Print Method for S3 Class "NRtest"

Prints the details of the NRtest object in a user-friendly manner. This method provides a clear and concise presentation of the test results contained within the NRtest object, including all relevant statistical metrics and test details.

Usage

## S3 method for class \pkg{NRtest}
# S3 method for class 'NRtest'
print(x, ...)

Arguments

x

an NRtest object.

...

further arguments passed to or from other methods.

Details

The print.NRtest function formats and presents the contents of the NRtest object, which includes statistical test results and related parameters. This function is designed to provide a user-friendly display of the object's contents, making it easier to understand the results of the analysis.

Value

Invisibly returns the input x.

See also

Author

Pengfei Wang nie23.wp8738@e.ntu.edu.sg